+++ packages/expo-localization/plugin/src/withExpoLocalization.ts
@@ -43,32 +43,49 @@ export function convertBcp47ToResourceQualifier(locale: string): string {
   return `b+${locale.replaceAll('-', '+')}`;
 }
 
+export function setAndroidSupportsRtl(
+  androidManifest: AndroidConfig.Manifest.AndroidManifest,
+  supportsRTL: boolean
+): AndroidConfig.Manifest.AndroidManifest {
+  const mainApplication = AndroidConfig.Manifest.getMainApplicationOrThrow(androidManifest);
+  mainApplication.$['android:supportsRtl'] = String(supportsRTL);
+  return androidManifest;
+}
+
 function withExpoLocalizationIos(config: ExpoConfig, data: ConfigPluginProps) {
-  const mergedConfig = { ...config.extra, ...data };
+  const {
+    supportsRTL,
+    forcesRTL,
+    supportedLocales: supportedLocalesOption,
+  } = {
+    ...config.extra,
+    ...data,
+  };
 
   const supportedLocales =
-    typeof mergedConfig.supportedLocales === 'object' &&
-    !Array.isArray(mergedConfig.supportedLocales)
-      ? mergedConfig.supportedLocales.ios
-      : mergedConfig.supportedLocales;
+    typeof supportedLocalesOption === 'object' && !Array.isArray(supportedLocalesOption)
+      ? supportedLocalesOption.ios
+      : supportedLocalesOption;
 
-  if (
-    mergedConfig?.supportsRTL == null &&
-    mergedConfig?.forcesRTL == null &&
-    supportedLocales == null
-  )
+  if (supportsRTL == null && forcesRTL == null && supportedLocales == null) {
     return config;
-  if (!config.ios) config.ios = {};
-  if (!config.ios.infoPlist) config.ios.infoPlist = {};
-  if (mergedConfig?.supportsRTL != null) {
-    config.ios.infoPlist.ExpoLocalization_supportsRTL = mergedConfig?.supportsRTL;
   }
-  if (mergedConfig?.forcesRTL != null) {
-    config.ios.infoPlist.ExpoLocalization_forcesRTL = mergedConfig?.forcesRTL;
+  if (!config.ios) {
+    config.ios = {};
+  }
+  if (!config.ios.infoPlist) {
+    config.ios.infoPlist = {};
+  }
+  if (supportsRTL != null) {
+    config.ios.infoPlist.ExpoLocalization_supportsRTL = supportsRTL;
+  }
+  if (forcesRTL != null) {
+    config.ios.infoPlist.ExpoLocalization_forcesRTL = forcesRTL;
   }
   if (supportedLocales != null) {
     config.ios.infoPlist.CFBundleLocalizations = supportedLocales;
   }
+
   return config;
 }
 
@@ -76,25 +93,42 @@ function withExpoLocalizationAndroid(config: ExpoConfig, data: ConfigPluginProps
   if (data.allowDynamicLocaleChangesAndroid) {
     config = withAndroidManifest(config, (config) => {
       const mainActivity = AndroidConfig.Manifest.getMainActivityOrThrow(config.modResults);
+
       if (!mainActivity.$['android:configChanges']?.includes('locale')) {
         mainActivity.$['android:configChanges'] += '|locale';
       }
       if (!mainActivity.$['android:configChanges']?.includes('layoutDirection')) {
         mainActivity.$['android:configChanges'] += '|layoutDirection';
       }
+
       return config;
     });
   }
-  const mergedConfig = { ...config.extra, ...data };
+
+  const {
+    supportsRTL,
+    forcesRTL,
+    supportedLocales: supportedLocalesOption,
+  } = {
+    ...config.extra,
+    ...data,
+  };
 
   const supportedLocales =
-    typeof mergedConfig.supportedLocales === 'object' &&
-    !Array.isArray(mergedConfig.supportedLocales)
-      ? mergedConfig.supportedLocales.android
-      : mergedConfig.supportedLocales;
+    typeof supportedLocalesOption === 'object' && !Array.isArray(supportedLocalesOption)
+      ? supportedLocalesOption.android
+      : supportedLocalesOption;
 
-  if (supportedLocales) {
+  if (supportsRTL != null) {
+    config = withAndroidManifest(config, (config) => {
+      config.modResults = setAndroidSupportsRtl(config.modResults, supportsRTL);
+      return config;
+    });
+  }
+
+  if (supportedLocales != null) {
     supportedLocales.forEach(assertLocale);
+
     config = withDangerousMod(config, [
       'android',
       (config) => {
@@ -102,6 +136,7 @@ function withExpoLocalizationAndroid(config: ExpoConfig, data: ConfigPluginProps
         const folder = path.join(projectRootPath, 'app/src/main/res/xml');
 
         fs.mkdirSync(folder, { recursive: true });
+
         fs.writeFileSync(
           path.join(folder, 'locales_config.xml'),
           [
@@ -115,6 +150,7 @@ function withExpoLocalizationAndroid(config: ExpoConfig, data: ConfigPluginProps
         return config;
       },
     ]);
+
     config = withAndroidManifest(config, (config) => {
       const mainApplication = AndroidConfig.Manifest.getMainApplicationOrThrow(config.modResults);
 
@@ -125,11 +161,13 @@ function withExpoLocalizationAndroid(config: ExpoConfig, data: ConfigPluginProps
 
       return config;
     });
+
     config = withAppBuildGradle(config, (config) => {
       if (config.modResults.language === 'groovy') {
         const resourceQualifiers = supportedLocales.map((locale) =>
           convertBcp47ToResourceQualifier(locale)
         );
+
         config.modResults.contents = AndroidConfig.CodeMod.appendContentsInsideDeclarationBlock(
           config.modResults.contents,
           'defaultConfig',
@@ -146,28 +184,29 @@ function withExpoLocalizationAndroid(config: ExpoConfig, data: ConfigPluginProps
     });
   }
   return withStringsXml(config, (config) => {
-    if (mergedConfig?.supportsRTL != null) {
+    if (supportsRTL != null) {
       config.modResults = AndroidConfig.Strings.setStringItem(
         [
           {
             $: { name: 'ExpoLocalization_supportsRTL', translatable: 'false' },
-            _: String(mergedConfig?.supportsRTL ?? 'unset'),
+            _: String(supportsRTL),
           },
         ],
         config.modResults
       );
     }
-    if (mergedConfig?.forcesRTL != null) {
+    if (forcesRTL != null) {
       config.modResults = AndroidConfig.Strings.setStringItem(
         [
           {
             $: { name: 'ExpoLocalization_forcesRTL', translatable: 'false' },
-            _: String(mergedConfig?.forcesRTL ?? 'unset'),
+            _: String(forcesRTL),
           },
         ],
         config.modResults
       );
     }
+
     return config;
   });
 }
@@ -178,6 +217,7 @@ function withExpoLocalization(config: ExpoConfig, data: ConfigPluginProps = {}) 
     ...data,
     allowDynamicLocaleChangesAndroid: data.allowDynamicLocaleChangesAndroid ?? true,
   };
+
   return withPlugins(config, [
     [withExpoLocalizationIos, normalizedData],
     [withExpoLocalizationAndroid, normalizedData],